Scenario for Huginn for Sauna automation
This json file is intended to be imported into Huginn- an IFTTT-like tool for doing your own automation.  It reads a custom arduino sensor I made for temperature and switch control that works with the AASER protocol and lightgw.py.  It calls oher services on the AASER network to change lights, but also texts me.